+++ fbw-a380x/src/systems/instruments/src/SD/Pages/Cruise/CruisePage.tsx
@@ -13,22 +13,26 @@ export const CruisePage = () => {
   const [unit] = usePersistentProperty('CONFIG_USING_METRIC_UNIT', '1');
   const [engine1FuelUsed] = useSimVar('L:A32NX_FUEL_USED:1', 'number', 1000);
   const [engine2FuelUsed] = useSimVar('L:A32NX_FUEL_USED:2', 'number', 1000);
+  const [engine3FuelUsed] = useSimVar('L:A32NX_FUEL_USED:3', 'number', 1000);
+  const [engine4FuelUsed] = useSimVar('L:A32NX_FUEL_USED:4', 'number', 1000);
 
   const engine1FuelUsedDisplay = fuelForDisplay(engine1FuelUsed, unit, 1, 5);
-  const engine2FuelUsedDisplay = engine1FuelUsedDisplay;
-  const engine3FuelUsedDisplay = fuelForDisplay(engine2FuelUsed, unit, 1, 5);
-  const engine4FuelUsedDisplay = engine3FuelUsedDisplay;
+  const engine2FuelUsedDisplay = fuelForDisplay(engine2FuelUsed, unit, 1, 5);
+  const engine3FuelUsedDisplay = fuelForDisplay(engine3FuelUsed, unit, 1, 5);
+  const engine4FuelUsedDisplay = fuelForDisplay(engine4FuelUsed, unit, 1, 5);
 
   const engineTotalFuelUsedDisplay =
     engine1FuelUsedDisplay + engine2FuelUsedDisplay + engine3FuelUsedDisplay + engine4FuelUsedDisplay;
 
   const [engine1FuelFlow] = useSimVar('L:A32NX_ENGINE_FF:1', 'number', 1000); // KG/HR
   const [engine2FuelFlow] = useSimVar('L:A32NX_ENGINE_FF:2', 'number', 1000);
+  const [engine3FuelFlow] = useSimVar('L:A32NX_ENGINE_FF:3', 'number', 1000);
+  const [engine4FuelFlow] = useSimVar('L:A32NX_ENGINE_FF:4', 'number', 1000);
 
   const engine1FuelFlowDisplay = fuelForDisplay(engine1FuelFlow, unit);
-  const engine2FuelFlowDisplay = engine1FuelFlowDisplay;
-  const engine3FuelFlowDisplay = fuelForDisplay(engine2FuelFlow, unit);
-  const engine4FuelFlowDisplay = engine3FuelFlowDisplay;
+  const engine2FuelFlowDisplay = fuelForDisplay(engine2FuelFlow, unit);
+  const engine3FuelFlowDisplay = fuelForDisplay(engine3FuelFlow, unit);
+  const engine4FuelFlowDisplay = fuelForDisplay(engine4FuelFlow, unit);
 
   // TODO Degraded accuracy indication for fuel flow and used