+++ .github/CHANGELOG.md
@@ -54,6 +54,7 @@
 1. [A380X/MFD] Add ATCCOM connect page - @heclak (heclak)
 1. [A380X/PFD] Fix CP VV button turning on VV on both PFDs - @heclak (Heclak)
 1. [A380X/FMS] Fixed handling of window altitude constraints, deletion of constraints, and formatting - @tracernz (Mike)
+1. [A380X/SD] Fix engine fuel flow and fuel used on cruise page still use only ENG 1 + 2 LVars @heclak (heclak)
 
 ## 0.12.0